+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix to setting Trade Policy Id
+
 ## [0.3.2] - 2022-02-07
 
 ### Changed